pyproject.toml

[tool.poetry]
name = "challenge-python-v1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "*"
psycopg2 = "*"
pytest = "*"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.5"
pillow = "^9.5.0"
factory-boy = "^3.2.1"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
